fastapi_docker_webapp/requirements.txt
accelerate==1.6.0
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
antlr4-python3-runtime==4.9.3
async-timeout==5.0.1
attrs==25.3.0
audioread==3.0.1
babel==2.17.0
bitsandbytes==0.45.5
boto3==1.37.37
botocore==1.37.37
cached_path==1.7.2
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.2
csvw==3.5.1
cycler==0.12.1
dacite==1.9.2
datasets==3.6.0
dill==0.3.8
distro==1.9.0
dlinfo==2.0.0
docker-pycreds==0.4.0
einops==0.8.1
einx==0.3.0
ema-pytorch==0.7.7
encodec==0.1.1
evaluate==0.4.5
exceptiongroup==1.2.2
fastapi==0.115.12
ffmpy==0.5.0
filelock==3.18.0
fire==0.7.1
fonttools==4.57.0
frozendict==2.4.6
frozenlist==1.6.0
fsspec==2024.12.0
gensim==4.3.3
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.2
google-auth==2.39.0
google-cloud-core==2.4.3
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
gradio==5.14.0
gradio_client==1.7.0
h11==0.14.0
htmlmin==0.1.12
httpcore==1.0.8
httpx==0.28.1
huggingface-hub==0.27.1
hydra-core==1.3.2
idna==3.10
ImageHash==4.3.1
ipywidgets==8.1.7
isodate==0.7.2
jieba==0.42.1
Jinja2==3.1.6
jiter==0.10.0
jiwer==4.0.0
jmespath==1.0.1
joblib==1.2.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyterlab_widgets==3.0.15
kiwisolver==1.4.8
langdetect==1.0.9
language-tags==1.2.0
lazy_loader==0.4
librosa==0.11.0
lightning==2.2.4
llvmlite==0.44.0
loguru==0.7.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.0
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.3
multimethod==1.12
multiprocess==0.70.16
networkx==3.4.2
nltk==3.9.1
numba==0.61.0
numpy==1.26.4
omegaconf==2.3.0
openai==1.90.0
orjson==3.10.16
pandas==2.2.3
patsy==1.0.1
peft==0.17.0
phik==0.12.4
phonemizer==3.3.0
pillow==11.2.1
platformdirs==4.3.7
pooch==1.8.2
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
puremagic==1.29
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
pydub==0.25.1
pyparsing==3.2.3
pypinyin==0.54.0
pythainlp==5.1.2
python-crfsuite==0.9.11
python-multipart==0.0.20
pytz==2025.2
PyWavelets==1.8.0
PyYAML==6.0.2
RapidFuzz==3.13.0
rdflib==7.1.4
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rfc3986==1.5.0
rich==13.9.4
rpds-py==0.27.1
rsa==4.9.1
ruff==0.11.6
s3transfer==0.11.5
safehttpx==0.1.6
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
segments==2.3.0
semantic-version==2.10.0
sentry-sdk==2.26.1
setproctitle==1.3.5
shellingham==1.5.4
sklearn-crfsuite==0.5.0
smart_open==7.3.1
smmap==5.0.2
soundfile==0.13.1
soxr==0.5.0.post1
ssg==0.0.8
starlette==0.46.2
statsmodels==0.14.4
syllapy==0.7.2
sympy==1.13.1
tabulate==0.9.0
termcolor==3.1.0
threadpoolctl==3.6.0
tltk==1.9.1
tokenizers==0.21.1
tomli==2.2.1
tomlkit==0.13.2
torchdiffeq==0.2.5
tqdm==4.67.1
transformers==4.50.3
transformers-stream-generator==0.0.5
typeguard==4.4.3
typer==0.15.2
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
uvicorn==0.34.2
visions==0.7.4
vocos==0.1.0
wandb==0.19.9
websockets==14.2
widgetsnbextension==4.0.14
win32_setctime==1.2.0
wordcloud==1.9.4
wrapt==1.17.3
x-transformers==2.2.12
xxhash==3.5.0
yarl==1.20.0
